packages/shared_lib/pyproject.toml
[project]
name = "shared-lib"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared_lib"]
